[DDW-794] Move to folder
