+++ source/renderer/app/components/voting/voting-info/VotingInfo.scss
@@ -1,4 +1,4 @@
-@import './votingConfig';
+@import '../votingConfig';
 
 .component {
   padding: 20px;
